use common SynthWorkload class
