+++ apps/OboeTester/app/src/main/cpp/cpu/AudioWorkloadTest.cpp
@@ -1,54 +1,5 @@
 #include "AudioWorkloadTest.h"
 
-void AudioWorkloadTest::SynthWorkload::onCallback(double workload) {
-    // If workload changes then restart notes.
-    if (workload != mPreviousWorkload) {
-        mSynth.allNotesOff();
-        mAreNotesOn = false;
-        mCountdown = 0; // trigger notes on
-        mPreviousWorkload = workload;
-    }
-    if (mCountdown <= 0) {
-        if (mAreNotesOn) {
-            mSynth.allNotesOff();
-            mAreNotesOn = false;
-            mCountdown = mOffFrames;
-        } else {
-            mSynth.notesOn((int)mPreviousWorkload);
-            mAreNotesOn = true;
-            mCountdown = mOnFrames;
-        }
-    }
-}
-
-/**
- * Render the notes into a stereo buffer.
- * Passing a nullptr will cause the calculated results to be discarded.
- * The workload should be the same.
- * @param buffer a real stereo buffer or nullptr
- * @param numFrames
- */
-void AudioWorkloadTest::SynthWorkload::renderStereo(float *buffer, int numFrames) {
-    if (buffer == nullptr) {
-        int framesLeft = numFrames;
-        while (framesLeft > 0) {
-            int framesThisTime = std::min(kDummyBufferSizeInFrames, framesLeft);
-            // Do the work then throw it away.
-            mSynth.renderStereo(mDummyStereoBuffer.get(), framesThisTime);
-            framesLeft -= framesThisTime;
-        }
-    } else {
-        mSynth.renderStereo(buffer, numFrames);
-    }
-    mCountdown -= numFrames;
-}
-
-AudioWorkloadTest::SynthWorkload::SynthWorkload(int onFrames, int offFrames) {
-    mSynth.setup(marksynth::kSynthmarkSampleRate, marksynth::kSynthmarkMaxVoices);
-    mOnFrames = onFrames;
-    mOffFrames = offFrames;
-}
-
 AudioWorkloadTest::AudioWorkloadTest() : mStream(nullptr) {}
 
 int32_t AudioWorkloadTest::open() {
